+++ intelmq/bots/collectors/rt/collector_rt.py
@@ -1,13 +1,12 @@
 # -*- coding: utf-8 -*-
-import io
 import re
-import zipfile
 from datetime import datetime, timedelta
 
 from dateutil import parser
 
 from intelmq.lib.bot import CollectorBot
-from intelmq.lib.utils import parse_relative, create_request_session_from_bot
+from intelmq.lib.utils import (parse_relative, create_request_session_from_bot,
+                               file_name_from_response, unzip)
 
 try:
     import rt
@@ -52,6 +51,18 @@ class RTCollectorBot(CollectorBot):
         self.set_request_parameters()
         self.session = create_request_session_from_bot(self)
 
+        if hasattr(self.parameters, 'unzip_attachment'):
+            self.logger.warning("The parameter 'unzip_attachment' is deprecated and "
+                                "will be removed in version 3.0 in favor of the "
+                                "more generic and powerful 'extract_files'. "
+                                "Look at the Bots documentation for more details.")
+            if not self.extract_files:
+                self.extract_files = self.parameters.unzip_attachment
+            else:
+                self.logger.warn("Both 'extract_files' and the deprecated "
+                                 "'unzip_attachment' parameter are in use. Ignoring "
+                                 "the latter one.")
+
     def process(self):
         RT = rt.Rt(self.parameters.uri, self.parameters.user,
                    self.parameters.password)
@@ -78,26 +89,27 @@ class RTCollectorBot(CollectorBot):
             ticket_id = int(ticket['id'].split('/')[1])
             self.logger.debug('Process ticket %s.', ticket_id)
             content = 'attachment'
-            for (att_id, att_name, _, _) in RT.get_attachments(ticket_id):
-                if not self.parameters.attachment_regex:
-                    break
-                if re.search(self.parameters.attachment_regex, att_name):
-                    self.logger.debug('Found attachment %s: %r.',
-                                      att_id, att_name)
-                    break
-            else:
-                urlmatch = False
-                if self.parameters.url_regex:
-                    ticket = RT.get_history(ticket_id)[0]
-                    created = ticket['Created']
-                    urlmatch = re.search(self.parameters.url_regex, ticket['Content'])
+            success = False
+            if self.parameters.attachment_regex:
+                for (att_id, att_name, _, _) in RT.get_attachments(ticket_id):
+                    if re.search(self.parameters.attachment_regex, att_name):
+                        self.logger.debug('Found attachment %s: %r.',
+                                          att_id, att_name)
+                        success = True
+                        content = 'attachment'
+                        break
+            if not success and self.parameters.url_regex:
+                ticket = RT.get_history(ticket_id)[0]
+                created = ticket['Created']
+                urlmatch = re.search(self.parameters.url_regex, ticket['Content'])
                 if urlmatch:
                     content = 'url'
                     url = urlmatch.group(0)
-                    self.logger.info('Matching URL found %r.', url)
-                else:
-                    self.logger.debug('No matching attachment or URL found.')
-                    continue
+                    self.logger.debug('Matching URL found %r.', url)
+                    success = True
+            if not success:
+                self.logger.info('No matching attachment or URL found.')
+                continue
 
             report = self.new_report()
 
@@ -105,13 +117,7 @@ class RTCollectorBot(CollectorBot):
                 attachment = RT.get_attachment_content(ticket_id, att_id)
                 created = RT.get_attachment(ticket_id, att_id)['Created']
 
-                if self.parameters.unzip_attachment:
-                    file_obj = io.BytesIO(attachment)
-                    zipped = zipfile.ZipFile(file_obj)
-                    raw = zipped.read(zipped.namelist()[0])
-                    report["extra.file_name"] = zipped.namelist()[0]
-                else:
-                    raw = attachment
+                raw = attachment
             else:
                 resp = self.session.get(url=url)
 
@@ -132,19 +138,20 @@ class RTCollectorBot(CollectorBot):
                         self.logger.info('Skipping now.')
                         continue
                 self.logger.info("Report #%d downloaded.", ticket_id)
-                raw = resp.text
+                if self.extract_files:
+                    raw = resp.content
+                else:
+                    raw = resp.text
+                report["extra.file_name"] = file_name_from_response(resp)
 
-            report.add("raw", raw)
             report.add("rtir_id", ticket_id)
             report.add("time.observation", created + ' UTC', overwrite=True)
-
             """
             On RT 3.8 these fields are only available on the original ticket, not the
             first history element as in 4.4
             """
             if "Subject" not in ticket:
                 ticket = RT.get_ticket(ticket_id)
-
             report.add("extra.email_subject", ticket["Subject"])
             report.add("extra.ticket_subject", ticket["Subject"])
             report.add("extra.email_from", ','.join(ticket["Requestors"]))
@@ -153,7 +160,31 @@ class RTCollectorBot(CollectorBot):
             report.add("extra.ticket_status", ticket["Status"])
             report.add("extra.ticket_owner", ticket["Owner"])
 
-            self.send_message(report)
+
+            if self.extract_files:
+                try:
+                    unzipped = unzip(raw, self.extract_files,
+                                     return_names=True, logger=self.logger)
+                except ValueError:
+                    self.logger.error('Could not uncompress the file. Skipping for now.')
+                    continue
+                for file_name, raw_report in unzipped:
+                    """
+                    File name priority is:
+                        From the archive (zip, tar.gz)
+                        From the HTTP Response
+                        From the Attachment name
+                        For gz attachments, only the last options works
+                    """
+                    report_new = report.copy()
+                    report_new.add("raw", raw_report)
+                    report_new.add("extra.file_name", file_name, overwrite=True)
+                    if not "extra.file_name" in report_new and att_name.endswith('.gz'):
+                        report_new["extra.file_name"] = att_name[:-3]
+                    self.send_message(report_new)
+            else:
+                report.add("raw", raw)
+                self.send_message(report)
 
             if self.parameters.take_ticket:
                 try: